+++ src/FlabIt.Guardians/StringGuardiansExtension.cs
@@ -83,7 +83,7 @@ namespace FlabIt.Guardians
         [NotNull]
         public static string ThrowIfNullOrEmpty(
             [System.Diagnostics.CodeAnalysis.NotNull, CanBeNull, ValidatedNotNull] this string? argument,
-            [CanBeNull, InvokerParameterName] string? argumentName = null,
+            [CanBeNull, InvokerParameterName, CallerArgumentExpression("argument")] string? argumentName = null,
             [CanBeNull] string? message = null)
         {
             argument = argument.ThrowIfNull(argumentName, message);
@@ -114,7 +114,7 @@ namespace FlabIt.Guardians
         [NotNull]
         public static string ThrowIfNullOrEmptyOrWhitespace(
             [System.Diagnostics.CodeAnalysis.NotNull, CanBeNull, ValidatedNotNull] this string? argument,
-            [CanBeNull, InvokerParameterName] string? argumentName = null,
+            [CanBeNull, InvokerParameterName, CallerArgumentExpression("argument")] string? argumentName = null,
             [CanBeNull] string? message = null)
         {
             return argument
@@ -138,7 +138,7 @@ namespace FlabIt.Guardians
         [NotNull]
         public static string ThrowIfNullOrWhitespace(
             [System.Diagnostics.CodeAnalysis.NotNull, CanBeNull, ValidatedNotNull] this string? argument,
-            [CanBeNull, InvokerParameterName] string? argumentName = null,
+            [CanBeNull, InvokerParameterName, CallerArgumentExpression("argument")] string? argumentName = null,
             [CanBeNull] string? message = null)
         {
             argument = argument.ThrowIfNull(argumentName, message);
@@ -169,7 +169,7 @@ namespace FlabIt.Guardians
         public static string ThrowIfShorterThan(
             [System.Diagnostics.CodeAnalysis.NotNull, NotNull, ValidatedNotNull] this string argument,
             int length,
-            [CanBeNull, InvokerParameterName] string? argumentName = null,
+            [CanBeNull, InvokerParameterName, CallerArgumentExpression("argument")] string? argumentName = null,
             [CanBeNull] string? message = null)
         {
             argument = argument.ThrowIfNull(argumentName, message);
@@ -200,7 +200,7 @@ namespace FlabIt.Guardians
         public static string ThrowIfLargerThan(
             [System.Diagnostics.CodeAnalysis.NotNull, NotNull, ValidatedNotNull] this string argument,
             int length,
-            [CanBeNull, InvokerParameterName] string? argumentName = null,
+            [CanBeNull, InvokerParameterName, CallerArgumentExpression("argument")] string? argumentName = null,
             [CanBeNull] string? message = null)
         {
             argument = argument.ThrowIfNull(argumentName, message);