+++ tests/FlabIt.Guardians.Tests/StringGuardiansExtensionNullOrEmptyOrWhitespaceTest.cs
@@ -17,9 +17,9 @@ namespace FlabIt.Guardians.Tests
         [TestCaseSource(nameof(NullStringsTestValuesSource))]
         public void When_calling_ThrowIfNullOrEmptyOrWhitespace_with_null_values_exception_argumentName_and_message_should_match_default(string testValue)
         {
-            var defaultMessage = string.Format(CultureInfo.InvariantCulture, FlabIt.Guardians.Properties.Resources.Exception_ArgumentNullMessageWithParamName, DefaultArgumentName);
+            var defaultMessage = string.Format(CultureInfo.InvariantCulture, FlabIt.Guardians.Properties.Resources.Exception_ArgumentNullMessageWithParamName, nameof(testValue));
 
-            AssertThatExceptionParamNameAndMessageShouldMatchDefaultArgumentName<ArgumentNullException>(() => StringGuardiansExtension.ThrowIfNullOrEmptyOrWhitespace(testValue), defaultMessage);
+            AssertThatExceptionParamNameAndMessageShouldMatchDefaultArgumentName<ArgumentNullException>(() => StringGuardiansExtension.ThrowIfNullOrEmptyOrWhitespace(testValue), defaultMessage, nameof(testValue));
         }
 
         [TestCaseSource(nameof(NullStringsTestValuesSource))]
@@ -51,11 +51,11 @@ namespace FlabIt.Guardians.Tests
         [TestCaseSource(nameof(EmptyStringsTestValuesSource))]
         public void When_calling_ThrowIfNullOrEmptyOrWhitespace_with_empty_values_exception_argumentName_and_message_should_match_default(string testValue)
         {
-            testValue.ThrowIfNull(nameof(testValue));
+            testValue.ThrowIfNull();
 
-            var defaultMessage = string.Format(CultureInfo.InvariantCulture, FlabIt.Guardians.Properties.Resources.Exception_ArgumentOfTypeXEmptyMessageWithParamName, DefaultArgumentName, typeof(string).FullName, testValue.Length);
+            var defaultMessage = string.Format(CultureInfo.InvariantCulture, FlabIt.Guardians.Properties.Resources.Exception_ArgumentOfTypeXEmptyMessageWithParamName, nameof(testValue), typeof(string).FullName, testValue.Length);
 
-            AssertThatExceptionParamNameAndMessageShouldMatchDefaultArgumentName<ArgumentEmptyException>(() => StringGuardiansExtension.ThrowIfNullOrEmptyOrWhitespace(testValue), defaultMessage);
+            AssertThatExceptionParamNameAndMessageShouldMatchDefaultArgumentName<ArgumentEmptyException>(() => StringGuardiansExtension.ThrowIfNullOrEmptyOrWhitespace(testValue), defaultMessage, nameof(testValue));
         }
 
         [TestCaseSource(nameof(EmptyStringsTestValuesSource))]
@@ -87,11 +87,11 @@ namespace FlabIt.Guardians.Tests
         [TestCaseSource(nameof(OnlyWhitespaceStringsTestValuesSource))]
         public void When_calling_ThrowIfNullOrEmptyOrWhitespace_with_whitespace_values_exception_argumentName_and_message_should_match_default(string testValue)
         {
-            testValue.ThrowIfNull(nameof(testValue));
+            testValue.ThrowIfNull();
 
-            var defaultMessage = string.Format(CultureInfo.InvariantCulture, FlabIt.Guardians.Properties.Resources.Exception_ArgumentOfTypeXOnlyWhitespaceMessageWithParamName, DefaultArgumentName, typeof(string).FullName);
+            var defaultMessage = string.Format(CultureInfo.InvariantCulture, FlabIt.Guardians.Properties.Resources.Exception_ArgumentOfTypeXOnlyWhitespaceMessageWithParamName, nameof(testValue), typeof(string).FullName);
 
-            AssertThatExceptionParamNameAndMessageShouldMatchDefaultArgumentName<ArgumentWhitespaceException>(() => StringGuardiansExtension.ThrowIfNullOrEmptyOrWhitespace(testValue), defaultMessage);
+            AssertThatExceptionParamNameAndMessageShouldMatchDefaultArgumentName<ArgumentWhitespaceException>(() => StringGuardiansExtension.ThrowIfNullOrEmptyOrWhitespace(testValue), defaultMessage, nameof(testValue));
         }
 
         [TestCaseSource(nameof(OnlyWhitespaceStringsTestValuesSource))]