+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@
 
 * Targeting for .NET 6.0
 * Remove PassThrough... guardians as announced in v0.8
+* Automatically determine correct argument name for exception message in case none is given
 * Updated JetBrains.Annotations package to 2021.3.0
 
 ## v0.8